requirements.txt
streamlit
rembg
pillow
streamlit-cropper
onnxruntime